refactor(session): split sessionAtom loaders into flat siblings
Move the 1036-line loaders.ts into loaderShared (logger/store/FetchPageResult),
mergeSessions (merge + pagination primitives), importedHistoryPaging
(external-history date buckets) and sidebarLoad (roster generation, category
pages, single-flight coordinator); loaders.ts keeps the same exported surface
and recomposes __TESTS_ONLY. Each store-keyed WeakMap still lives in exactly
one module. Pure code motion, no behaviour change.
